fix: build with `version.dev=no`
This commit fixes a minor issue where Terraform was being built
with `version.dev=true`, resulting in the embedded version string
including the `-dev` prefix, and also triggering a warning about
the version being out of date.

The fix is to use Go's LDFLAGS to set `version.dev=no` at build
time, which is also done in the upstream repository.

Fixes #130